move guidance tag loader tests
