+++ api-js/src/loaders/guidance-tags/__tests__/load-ssl-guidance-tags-connections.test.js
@@ -5,16 +5,16 @@ const { ArangoTools, dbNameFromFile } = require('arango-tools')
 const { toGlobalId } = require('graphql-relay')
 const { setupI18n } = require('@lingui/core')
 
-const englishMessages = require('../locale/en/messages')
-const frenchMessages = require('../locale/fr/messages')
-const { makeMigrations } = require('../../migrations')
-const { cleanseInput } = require('../validators')
+const englishMessages = require('../../../locale/en/messages')
+const frenchMessages = require('../../../locale/fr/messages')
+const { makeMigrations } = require('../../../../migrations')
+const { cleanseInput } = require('../../../validators')
 const {
-  dkimGuidanceTagConnectionsLoader,
-  dkimGuidanceTagLoader,
-} = require('../loaders')
+  sslGuidanceTagConnectionsLoader,
+  sslGuidanceTagLoader,
+} = require('../..')
 
-describe('when given the load dkim guidance tag connection function', () => {
+describe('when given the load ssl guidance tag connection function', () => {
   let query, drop, truncate, migrate, collections, user, i18n
 
   const consoleWarnOutput = []
@@ -54,11 +54,11 @@ describe('when given the load dkim guidance tag connection function', () => {
       emailValidated: false,
     })
 
-    await collections.dkimGuidanceTags.save({
-      _key: 'dkim1',
+    await collections.sslGuidanceTags.save({
+      _key: 'ssl1',
     })
-    await collections.dkimGuidanceTags.save({
-      _key: 'dkim2',
+    await collections.sslGuidanceTags.save({
+      _key: 'ssl2',
     })
   })
 
@@ -68,39 +68,39 @@ describe('when given the load dkim guidance tag connection function', () => {
 
   describe('given a successful load', () => {
     describe('using no cursor', () => {
-      it('returns multiple dkim results', async () => {
-        const connectionLoader = dkimGuidanceTagConnectionsLoader(
+      it('returns multiple ssl results', async () => {
+        const connectionLoader = sslGuidanceTagConnectionsLoader(
           query,
           user._key,
           cleanseInput,
           i18n,
         )
 
-        const dkimGuidanceTags = ['dkim1', 'dkim2']
+        const sslGuidanceTags = ['ssl1', 'ssl2']
         const connectionArgs = {
           first: 5,
         }
 
-        const dkimTags = await connectionLoader({
-          dkimGuidanceTags,
+        const sslTags = await connectionLoader({
+          sslGuidanceTags,
           ...connectionArgs,
         })
 
-        const dkimTagLoader = dkimGuidanceTagLoader(query)
-        const expectedDkimTags = await dkimTagLoader.loadMany(dkimGuidanceTags)
+        const sslTagLoader = sslGuidanceTagLoader(query)
+        const expectedSslTags = await sslTagLoader.loadMany(sslGuidanceTags)
 
         const expectedStructure = {
           edges: [
             {
-              cursor: toGlobalId('guidanceTags', expectedDkimTags[0]._key),
+              cursor: toGlobalId('guidanceTags', expectedSslTags[0]._key),
               node: {
-                ...expectedDkimTags[0],
+                ...expectedSslTags[0],
               },
             },
             {
-              cursor: toGlobalId('guidanceTags', expectedDkimTags[1]._key),
+              cursor: toGlobalId('guidanceTags', expectedSslTags[1]._key),
               node: {
-                ...expectedDkimTags[1],
+                ...expectedSslTags[1],
               },
             },
           ],
@@ -108,44 +108,44 @@ describe('when given the load dkim guidance tag connection function', () => {
           pageInfo: {
             hasNextPage: false,
             hasPreviousPage: false,
-            startCursor: toGlobalId('guidanceTags', expectedDkimTags[0]._key),
-            endCursor: toGlobalId('guidanceTags', expectedDkimTags[1]._key),
+            startCursor: toGlobalId('guidanceTags', expectedSslTags[0]._key),
+            endCursor: toGlobalId('guidanceTags', expectedSslTags[1]._key),
           },
         }
 
-        expect(dkimTags).toEqual(expectedStructure)
+        expect(sslTags).toEqual(expectedStructure)
       })
     })
     describe('using after cursor', () => {
-      it('returns dkim result(s) after a given node id', async () => {
-        const connectionLoader = dkimGuidanceTagConnectionsLoader(
+      it('returns ssl result(s) after a given node id', async () => {
+        const connectionLoader = sslGuidanceTagConnectionsLoader(
           query,
           user._key,
           cleanseInput,
           i18n,
         )
 
-        const dkimGuidanceTags = ['dkim1', 'dkim2']
+        const sslGuidanceTags = ['ssl1', 'ssl2']
 
-        const dkimTagLoader = dkimGuidanceTagLoader(query)
-        const expectedDkimTags = await dkimTagLoader.loadMany(dkimGuidanceTags)
+        const sslTagLoader = sslGuidanceTagLoader(query)
+        const expectedSslTags = await sslTagLoader.loadMany(sslGuidanceTags)
 
         const connectionArgs = {
           first: 5,
-          after: toGlobalId('guidanceTag', expectedDkimTags[0]._key),
+          after: toGlobalId('guidanceTag', expectedSslTags[0]._key),
         }
 
-        const dkimTags = await connectionLoader({
-          dkimGuidanceTags,
+        const sslTags = await connectionLoader({
+          sslGuidanceTags,
           ...connectionArgs,
         })
 
         const expectedStructure = {
           edges: [
             {
-              cursor: toGlobalId('guidanceTags', expectedDkimTags[1]._key),
+              cursor: toGlobalId('guidanceTags', expectedSslTags[1]._key),
               node: {
-                ...expectedDkimTags[1],
+                ...expectedSslTags[1],
               },
             },
           ],
@@ -153,44 +153,44 @@ describe('when given the load dkim guidance tag connection function', () => {
           pageInfo: {
             hasNextPage: false,
             hasPreviousPage: true,
-            startCursor: toGlobalId('guidanceTags', expectedDkimTags[1]._key),
-            endCursor: toGlobalId('guidanceTags', expectedDkimTags[1]._key),
+            startCursor: toGlobalId('guidanceTags', expectedSslTags[1]._key),
+            endCursor: toGlobalId('guidanceTags', expectedSslTags[1]._key),
           },
         }
 
-        expect(dkimTags).toEqual(expectedStructure)
+        expect(sslTags).toEqual(expectedStructure)
       })
     })
     describe('using before cursor', () => {
-      it('returns dkim result(s) before a given node id', async () => {
-        const connectionLoader = dkimGuidanceTagConnectionsLoader(
+      it('returns ssl result(s) before a given node id', async () => {
+        const connectionLoader = sslGuidanceTagConnectionsLoader(
           query,
           user._key,
           cleanseInput,
           i18n,
         )
 
-        const dkimGuidanceTags = ['dkim1', 'dkim2']
+        const sslGuidanceTags = ['ssl1', 'ssl2']
 
-        const dkimTagLoader = dkimGuidanceTagLoader(query)
-        const expectedDkimTags = await dkimTagLoader.loadMany(dkimGuidanceTags)
+        const sslTagLoader = sslGuidanceTagLoader(query)
+        const expectedSslTags = await sslTagLoader.loadMany(sslGuidanceTags)
 
         const connectionArgs = {
           first: 5,
-          before: toGlobalId('guidanceTags', expectedDkimTags[1]._key),
+          before: toGlobalId('guidanceTags', expectedSslTags[1]._key),
         }
 
-        const dkimTags = await connectionLoader({
-          dkimGuidanceTags,
+        const sslTags = await connectionLoader({
+          sslGuidanceTags,
           ...connectionArgs,
         })
 
         const expectedStructure = {
           edges: [
             {
-              cursor: toGlobalId('guidanceTags', expectedDkimTags[0]._key),
+              cursor: toGlobalId('guidanceTags', expectedSslTags[0]._key),
               node: {
-                ...expectedDkimTags[0],
+                ...expectedSslTags[0],
               },
             },
           ],
@@ -198,43 +198,43 @@ describe('when given the load dkim guidance tag connection function', () => {
           pageInfo: {
             hasNextPage: true,
             hasPreviousPage: false,
-            startCursor: toGlobalId('guidanceTags', expectedDkimTags[0]._key),
-            endCursor: toGlobalId('guidanceTags', expectedDkimTags[0]._key),
+            startCursor: toGlobalId('guidanceTags', expectedSslTags[0]._key),
+            endCursor: toGlobalId('guidanceTags', expectedSslTags[0]._key),
           },
         }
 
-        expect(dkimTags).toEqual(expectedStructure)
+        expect(sslTags).toEqual(expectedStructure)
       })
     })
     describe('using first limit', () => {
       it('returns the first n amount of item(s)', async () => {
-        const connectionLoader = dkimGuidanceTagConnectionsLoader(
+        const connectionLoader = sslGuidanceTagConnectionsLoader(
           query,
           user._key,
           cleanseInput,
           i18n,
         )
 
-        const dkimGuidanceTags = ['dkim1', 'dkim2']
+        const sslGuidanceTags = ['ssl1', 'ssl2']
 
-        const dkimTagLoader = dkimGuidanceTagLoader(query)
-        const expectedDkimTags = await dkimTagLoader.loadMany(dkimGuidanceTags)
+        const sslTagLoader = sslGuidanceTagLoader(query)
+        const expectedSslTags = await sslTagLoader.loadMany(sslGuidanceTags)
 
         const connectionArgs = {
           first: 1,
         }
 
-        const dkimTags = await connectionLoader({
-          dkimGuidanceTags,
+        const sslTags = await connectionLoader({
+          sslGuidanceTags,
           ...connectionArgs,
         })
 
         const expectedStructure = {
           edges: [
             {
-              cursor: toGlobalId('guidanceTags', expectedDkimTags[0]._key),
+              cursor: toGlobalId('guidanceTags', expectedSslTags[0]._key),
               node: {
-                ...expectedDkimTags[0],
+                ...expectedSslTags[0],
               },
             },
           ],
@@ -242,43 +242,43 @@ describe('when given the load dkim guidance tag connection function', () => {
           pageInfo: {
             hasNextPage: true,
             hasPreviousPage: false,
-            startCursor: toGlobalId('guidanceTags', expectedDkimTags[0]._key),
-            endCursor: toGlobalId('guidanceTags', expectedDkimTags[0]._key),
+            startCursor: toGlobalId('guidanceTags', expectedSslTags[0]._key),
+            endCursor: toGlobalId('guidanceTags', expectedSslTags[0]._key),
           },
         }
 
-        expect(dkimTags).toEqual(expectedStructure)
+        expect(sslTags).toEqual(expectedStructure)
       })
     })
     describe('using last limit', () => {
       it('returns the last n amount of item(s)', async () => {
-        const connectionLoader = dkimGuidanceTagConnectionsLoader(
+        const connectionLoader = sslGuidanceTagConnectionsLoader(
           query,
           user._key,
           cleanseInput,
           i18n,
         )
 
-        const dkimGuidanceTags = ['dkim1', 'dkim2']
+        const sslGuidanceTags = ['ssl1', 'ssl2']
 
-        const dkimTagLoader = dkimGuidanceTagLoader(query)
-        const expectedDkimTags = await dkimTagLoader.loadMany(dkimGuidanceTags)
+        const sslTagLoader = sslGuidanceTagLoader(query)
+        const expectedSslTags = await sslTagLoader.loadMany(sslGuidanceTags)
 
         const connectionArgs = {
           last: 1,
         }
 
-        const dkimTags = await connectionLoader({
-          dkimGuidanceTags,
+        const sslTags = await connectionLoader({
+          sslGuidanceTags,
           ...connectionArgs,
         })
 
         const expectedStructure = {
           edges: [
             {
-              cursor: toGlobalId('guidanceTags', expectedDkimTags[1]._key),
+              cursor: toGlobalId('guidanceTags', expectedSslTags[1]._key),
               node: {
-                ...expectedDkimTags[1],
+                ...expectedSslTags[1],
               },
             },
           ],
@@ -286,18 +286,18 @@ describe('when given the load dkim guidance tag connection function', () => {
           pageInfo: {
             hasNextPage: false,
             hasPreviousPage: true,
-            startCursor: toGlobalId('guidanceTags', expectedDkimTags[1]._key),
-            endCursor: toGlobalId('guidanceTags', expectedDkimTags[1]._key),
+            startCursor: toGlobalId('guidanceTags', expectedSslTags[1]._key),
+            endCursor: toGlobalId('guidanceTags', expectedSslTags[1]._key),
           },
         }
 
-        expect(dkimTags).toEqual(expectedStructure)
+        expect(sslTags).toEqual(expectedStructure)
       })
     })
-    describe('no dkim results are found', () => {
+    describe('no ssl results are found', () => {
       it('returns an empty structure', async () => {
         await truncate()
-        const connectionLoader = dkimGuidanceTagConnectionsLoader(
+        const connectionLoader = sslGuidanceTagConnectionsLoader(
           query,
           user._key,
           cleanseInput,
@@ -308,9 +308,9 @@ describe('when given the load dkim guidance tag connection function', () => {
           first: 5,
         }
 
-        const dkimGuidanceTags = ['dkim1', 'dkim2']
-        const dkimTags = await connectionLoader({
-          dkimGuidanceTags,
+        const sslGuidanceTags = ['ssl1', 'ssl2']
+        const sslTags = await connectionLoader({
+          sslGuidanceTags,
           ...connectionArgs,
         })
 
@@ -325,7 +325,7 @@ describe('when given the load dkim guidance tag connection function', () => {
           },
         }
 
-        expect(dkimTags).toEqual(expectedStructure)
+        expect(sslTags).toEqual(expectedStructure)
       })
     })
   })
@@ -344,19 +344,19 @@ describe('when given the load dkim guidance tag connection function', () => {
     describe('given a unsuccessful load', () => {
       describe('both limits are not set', () => {
         it('returns an error message', async () => {
-          const connectionLoader = dkimGuidanceTagConnectionsLoader(
+          const connectionLoader = sslGuidanceTagConnectionsLoader(
             query,
             user._key,
             cleanseInput,
             i18n,
           )
 
-          const dkimGuidanceTags = ['dkim1', 'dkim2']
+          const sslGuidanceTags = ['ssl1', 'ssl2']
           const connectionArgs = {}
 
           try {
             await connectionLoader({
-              dkimGuidanceTags,
+              sslGuidanceTags,
               ...connectionArgs,
             })
           } catch (err) {
@@ -367,20 +367,20 @@ describe('when given the load dkim guidance tag connection function', () => {
             )
           }
           expect(consoleWarnOutput).toEqual([
-            `User: ${user._key} did not have either \`first\` or \`last\` arguments set for: dkimGuidanceTagConnectionsLoader.`,
+            `User: ${user._key} did not have either \`first\` or \`last\` arguments set for: sslGuidanceTagConnectionsLoader.`,
           ])
         })
       })
       describe('both limits are set', () => {
         it('returns an error message', async () => {
-          const connectionLoader = dkimGuidanceTagConnectionsLoader(
+          const connectionLoader = sslGuidanceTagConnectionsLoader(
             query,
             user._key,
             cleanseInput,
             i18n,
           )
 
-          const dkimGuidanceTags = ['dkim1', 'dkim2']
+          const sslGuidanceTags = ['ssl1', 'ssl2']
           const connectionArgs = {
             first: 1,
             last: 5,
@@ -388,7 +388,7 @@ describe('when given the load dkim guidance tag connection function', () => {
 
           try {
             await connectionLoader({
-              dkimGuidanceTags,
+              sslGuidanceTags,
               ...connectionArgs,
             })
           } catch (err) {
@@ -399,28 +399,28 @@ describe('when given the load dkim guidance tag connection function', () => {
             )
           }
           expect(consoleWarnOutput).toEqual([
-            `User: ${user._key} attempted to have \`first\` and \`last\` arguments set for: dkimGuidanceTagConnectionsLoader.`,
+            `User: ${user._key} attempted to have \`first\` and \`last\` arguments set for: sslGuidanceTagConnectionsLoader.`,
           ])
         })
       })
       describe('limits are below minimum', () => {
         describe('first is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dkimGuidanceTagConnectionsLoader(
+            const connectionLoader = sslGuidanceTagConnectionsLoader(
               query,
               user._key,
               cleanseInput,
               i18n,
             )
 
-            const dkimGuidanceTags = ['dkim1', 'dkim2']
+            const sslGuidanceTags = ['ssl1', 'ssl2']
             const connectionArgs = {
               first: -1,
             }
 
             try {
               await connectionLoader({
-                dkimGuidanceTags,
+                sslGuidanceTags,
                 ...connectionArgs,
               })
             } catch (err) {
@@ -431,27 +431,27 @@ describe('when given the load dkim guidance tag connection function', () => {
               )
             }
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`first\` set below zero for: dkimGuidanceTagConnectionsLoader.`,
+              `User: ${user._key} attempted to have \`first\` set below zero for: sslGuidanceTagConnectionsLoader.`,
             ])
           })
         })
         describe('last is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dkimGuidanceTagConnectionsLoader(
+            const connectionLoader = sslGuidanceTagConnectionsLoader(
               query,
               user._key,
               cleanseInput,
               i18n,
             )
 
-            const dkimGuidanceTags = ['dkim1', 'dkim2']
+            const sslGuidanceTags = ['ssl1', 'ssl2']
             const connectionArgs = {
               last: -5,
             }
 
             try {
               await connectionLoader({
-                dkimGuidanceTags,
+                sslGuidanceTags,
                 ...connectionArgs,
               })
             } catch (err) {
@@ -462,7 +462,7 @@ describe('when given the load dkim guidance tag connection function', () => {
               )
             }
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`last\` set below zero for: dkimGuidanceTagConnectionsLoader.`,
+              `User: ${user._key} attempted to have \`last\` set below zero for: sslGuidanceTagConnectionsLoader.`,
             ])
           })
         })
@@ -470,21 +470,21 @@ describe('when given the load dkim guidance tag connection function', () => {
       describe('limits are above maximum', () => {
         describe('first is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dkimGuidanceTagConnectionsLoader(
+            const connectionLoader = sslGuidanceTagConnectionsLoader(
               query,
               user._key,
               cleanseInput,
               i18n,
             )
 
-            const dkimGuidanceTags = ['dkim1', 'dkim2']
+            const sslGuidanceTags = ['ssl1', 'ssl2']
             const connectionArgs = {
               first: 1000,
             }
 
             try {
               await connectionLoader({
-                dkimGuidanceTags,
+                sslGuidanceTags,
                 ...connectionArgs,
               })
             } catch (err) {
@@ -495,27 +495,27 @@ describe('when given the load dkim guidance tag connection function', () => {
               )
             }
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`first\` set to 1000 for: dkimGuidanceTagConnectionsLoader.`,
+              `User: ${user._key} attempted to have \`first\` set to 1000 for: sslGuidanceTagConnectionsLoader.`,
             ])
           })
         })
         describe('last is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dkimGuidanceTagConnectionsLoader(
+            const connectionLoader = sslGuidanceTagConnectionsLoader(
               query,
               user._key,
               cleanseInput,
               i18n,
             )
 
-            const dkimGuidanceTags = ['dkim1', 'dkim2']
+            const sslGuidanceTags = ['ssl1', 'ssl2']
             const connectionArgs = {
               last: 500,
             }
 
             try {
               await connectionLoader({
-                dkimGuidanceTags,
+                sslGuidanceTags,
                 ...connectionArgs,
               })
             } catch (err) {
@@ -526,7 +526,7 @@ describe('when given the load dkim guidance tag connection function', () => {
               )
             }
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`last\` set to 500 for: dkimGuidanceTagConnectionsLoader.`,
+              `User: ${user._key} attempted to have \`last\` set to 500 for: sslGuidanceTagConnectionsLoader.`,
             ])
           })
         })
@@ -537,21 +537,21 @@ describe('when given the load dkim guidance tag connection function', () => {
             it(`returns an error when first set to ${stringify(
               invalidInput,
             )}`, async () => {
-              const connectionLoader = dkimGuidanceTagConnectionsLoader(
+              const connectionLoader = sslGuidanceTagConnectionsLoader(
                 query,
                 user._key,
                 cleanseInput,
                 i18n,
               )
 
-              const dkimGuidanceTags = ['dkim1', 'dkim2']
+              const sslGuidanceTags = ['ssl1', 'ssl2']
               const connectionArgs = {
                 first: invalidInput,
               }
 
               try {
                 await connectionLoader({
-                  dkimGuidanceTags,
+                  sslGuidanceTags,
                   ...connectionArgs,
                 })
               } catch (err) {
@@ -564,7 +564,7 @@ describe('when given the load dkim guidance tag connection function', () => {
               expect(consoleWarnOutput).toEqual([
                 `User: ${
                   user._key
-                } attempted to have \`first\` set as a ${typeof invalidInput} for: dkimGuidanceTagConnectionsLoader.`,
+                } attempted to have \`first\` set as a ${typeof invalidInput} for: sslGuidanceTagConnectionsLoader.`,
               ])
             })
           })
@@ -574,21 +574,21 @@ describe('when given the load dkim guidance tag connection function', () => {
             it(`returns an error when last set to ${stringify(
               invalidInput,
             )}`, async () => {
-              const connectionLoader = dkimGuidanceTagConnectionsLoader(
+              const connectionLoader = sslGuidanceTagConnectionsLoader(
                 query,
                 user._key,
                 cleanseInput,
                 i18n,
               )
 
-              const dkimGuidanceTags = ['dkim1', 'dkim2']
+              const sslGuidanceTags = ['ssl1', 'ssl2']
               const connectionArgs = {
                 last: invalidInput,
               }
 
               try {
                 await connectionLoader({
-                  dkimGuidanceTags,
+                  sslGuidanceTags,
                   ...connectionArgs,
                 })
               } catch (err) {
@@ -601,7 +601,7 @@ describe('when given the load dkim guidance tag connection function', () => {
               expect(consoleWarnOutput).toEqual([
                 `User: ${
                   user._key
-                } attempted to have \`last\` set as a ${typeof invalidInput} for: dkimGuidanceTagConnectionsLoader.`,
+                } attempted to have \`last\` set as a ${typeof invalidInput} for: sslGuidanceTagConnectionsLoader.`,
               ])
             })
           })
@@ -614,30 +614,30 @@ describe('when given the load dkim guidance tag connection function', () => {
           .fn()
           .mockRejectedValue(new Error('Database Error Occurred.'))
 
-        const connectionLoader = dkimGuidanceTagConnectionsLoader(
+        const connectionLoader = sslGuidanceTagConnectionsLoader(
           query,
           user._key,
           cleanseInput,
           i18n,
         )
 
-        const dkimGuidanceTags = ['dkim1', 'dkim2']
+        const sslGuidanceTags = ['ssl1', 'ssl2']
         const connectionArgs = {
           first: 5,
         }
         try {
           await connectionLoader({
-            dkimGuidanceTags,
+            sslGuidanceTags,
             ...connectionArgs,
           })
         } catch (err) {
           expect(err).toEqual(
-            new Error('Unable to load dkim guidance tags. Please try again.'),
+            new Error('Unable to load ssl guidance tags. Please try again.'),
           )
         }
 
         expect(consoleErrorOutput).toEqual([
-          `Database error occurred while user: ${user._key} was trying to gather orgs in dkimGuidanceTagConnectionsLoader, error: Error: Database Error Occurred.`,
+          `Database error occurred while user: ${user._key} was trying to gather orgs in sslGuidanceTagConnectionsLoader, error: Error: Database Error Occurred.`,
         ])
       })
     })
@@ -650,30 +650,30 @@ describe('when given the load dkim guidance tag connection function', () => {
         }
         const query = jest.fn().mockReturnValueOnce(cursor)
 
-        const connectionLoader = dkimGuidanceTagConnectionsLoader(
+        const connectionLoader = sslGuidanceTagConnectionsLoader(
           query,
           user._key,
           cleanseInput,
           i18n,
         )
 
-        const dkimGuidanceTags = ['dkim1', 'dkim2']
+        const sslGuidanceTags = ['ssl1', 'ssl2']
         const connectionArgs = {
           first: 5,
         }
         try {
           await connectionLoader({
-            dkimGuidanceTags,
+            sslGuidanceTags,
             ...connectionArgs,
           })
         } catch (err) {
           expect(err).toEqual(
-            new Error('Unable to load dkim guidance tags. Please try again.'),
+            new Error('Unable to load ssl guidance tags. Please try again.'),
           )
         }
 
         expect(consoleErrorOutput).toEqual([
-          `Cursor error occurred while user: ${user._key} was trying to gather orgs in dkimGuidanceTagConnectionsLoader, error: Error: Cursor Error Occurred.`,
+          `Cursor error occurred while user: ${user._key} was trying to gather orgs in sslGuidanceTagConnectionsLoader, error: Error: Cursor Error Occurred.`,
         ])
       })
     })
@@ -693,39 +693,39 @@ describe('when given the load dkim guidance tag connection function', () => {
     describe('given a unsuccessful load', () => {
       describe('both limits are not set', () => {
         it('returns an error message', async () => {
-          const connectionLoader = dkimGuidanceTagConnectionsLoader(
+          const connectionLoader = sslGuidanceTagConnectionsLoader(
             query,
             user._key,
             cleanseInput,
             i18n,
           )
 
-          const dkimGuidanceTags = ['dkim1', 'dkim2']
+          const sslGuidanceTags = ['ssl1', 'ssl2']
           const connectionArgs = {}
 
           try {
             await connectionLoader({
-              dkimGuidanceTags,
+              sslGuidanceTags,
               ...connectionArgs,
             })
           } catch (err) {
             expect(err).toEqual(new Error('todo'))
           }
           expect(consoleWarnOutput).toEqual([
-            `User: ${user._key} did not have either \`first\` or \`last\` arguments set for: dkimGuidanceTagConnectionsLoader.`,
+            `User: ${user._key} did not have either \`first\` or \`last\` arguments set for: sslGuidanceTagConnectionsLoader.`,
           ])
         })
       })
       describe('both limits are set', () => {
         it('returns an error message', async () => {
-          const connectionLoader = dkimGuidanceTagConnectionsLoader(
+          const connectionLoader = sslGuidanceTagConnectionsLoader(
             query,
             user._key,
             cleanseInput,
             i18n,
           )
 
-          const dkimGuidanceTags = ['dkim1', 'dkim2']
+          const sslGuidanceTags = ['ssl1', 'ssl2']
           const connectionArgs = {
             first: 1,
             last: 5,
@@ -733,69 +733,69 @@ describe('when given the load dkim guidance tag connection function', () => {
 
           try {
             await connectionLoader({
-              dkimGuidanceTags,
+              sslGuidanceTags,
               ...connectionArgs,
             })
           } catch (err) {
             expect(err).toEqual(new Error('todo'))
           }
           expect(consoleWarnOutput).toEqual([
-            `User: ${user._key} attempted to have \`first\` and \`last\` arguments set for: dkimGuidanceTagConnectionsLoader.`,
+            `User: ${user._key} attempted to have \`first\` and \`last\` arguments set for: sslGuidanceTagConnectionsLoader.`,
           ])
         })
       })
       describe('limits are below minimum', () => {
         describe('first is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dkimGuidanceTagConnectionsLoader(
+            const connectionLoader = sslGuidanceTagConnectionsLoader(
               query,
               user._key,
               cleanseInput,
               i18n,
             )
 
-            const dkimGuidanceTags = ['dkim1', 'dkim2']
+            const sslGuidanceTags = ['ssl1', 'ssl2']
             const connectionArgs = {
               first: -1,
             }
 
             try {
               await connectionLoader({
-                dkimGuidanceTags,
+                sslGuidanceTags,
                 ...connectionArgs,
               })
             } catch (err) {
               expect(err).toEqual(new Error('todo'))
             }
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`first\` set below zero for: dkimGuidanceTagConnectionsLoader.`,
+              `User: ${user._key} attempted to have \`first\` set below zero for: sslGuidanceTagConnectionsLoader.`,
             ])
           })
         })
         describe('last is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dkimGuidanceTagConnectionsLoader(
+            const connectionLoader = sslGuidanceTagConnectionsLoader(
               query,
               user._key,
               cleanseInput,
               i18n,
             )
 
-            const dkimGuidanceTags = ['dkim1', 'dkim2']
+            const sslGuidanceTags = ['ssl1', 'ssl2']
             const connectionArgs = {
               last: -5,
             }
 
             try {
               await connectionLoader({
-                dkimGuidanceTags,
+                sslGuidanceTags,
                 ...connectionArgs,
               })
             } catch (err) {
               expect(err).toEqual(new Error('todo'))
             }
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`last\` set below zero for: dkimGuidanceTagConnectionsLoader.`,
+              `User: ${user._key} attempted to have \`last\` set below zero for: sslGuidanceTagConnectionsLoader.`,
             ])
           })
         })
@@ -803,55 +803,55 @@ describe('when given the load dkim guidance tag connection function', () => {
       describe('limits are above maximum', () => {
         describe('first is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dkimGuidanceTagConnectionsLoader(
+            const connectionLoader = sslGuidanceTagConnectionsLoader(
               query,
               user._key,
               cleanseInput,
               i18n,
             )
 
-            const dkimGuidanceTags = ['dkim1', 'dkim2']
+            const sslGuidanceTags = ['ssl1', 'ssl2']
             const connectionArgs = {
               first: 1000,
             }
 
             try {
               await connectionLoader({
-                dkimGuidanceTags,
+                sslGuidanceTags,
                 ...connectionArgs,
               })
             } catch (err) {
               expect(err).toEqual(new Error('todo'))
             }
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`first\` set to 1000 for: dkimGuidanceTagConnectionsLoader.`,
+              `User: ${user._key} attempted to have \`first\` set to 1000 for: sslGuidanceTagConnectionsLoader.`,
             ])
           })
         })
         describe('last is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dkimGuidanceTagConnectionsLoader(
+            const connectionLoader = sslGuidanceTagConnectionsLoader(
               query,
               user._key,
               cleanseInput,
               i18n,
             )
 
-            const dkimGuidanceTags = ['dkim1', 'dkim2']
+            const sslGuidanceTags = ['ssl1', 'ssl2']
             const connectionArgs = {
               last: 500,
             }
 
             try {
               await connectionLoader({
-                dkimGuidanceTags,
+                sslGuidanceTags,
                 ...connectionArgs,
               })
             } catch (err) {
               expect(err).toEqual(new Error('todo'))
             }
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`last\` set to 500 for: dkimGuidanceTagConnectionsLoader.`,
+              `User: ${user._key} attempted to have \`last\` set to 500 for: sslGuidanceTagConnectionsLoader.`,
             ])
           })
         })
@@ -862,21 +862,21 @@ describe('when given the load dkim guidance tag connection function', () => {
             it(`returns an error when first set to ${stringify(
               invalidInput,
             )}`, async () => {
-              const connectionLoader = dkimGuidanceTagConnectionsLoader(
+              const connectionLoader = sslGuidanceTagConnectionsLoader(
                 query,
                 user._key,
                 cleanseInput,
                 i18n,
               )
 
-              const dkimGuidanceTags = ['dkim1', 'dkim2']
+              const sslGuidanceTags = ['ssl1', 'ssl2']
               const connectionArgs = {
                 first: invalidInput,
               }
 
               try {
                 await connectionLoader({
-                  dkimGuidanceTags,
+                  sslGuidanceTags,
                   ...connectionArgs,
                 })
               } catch (err) {
@@ -885,7 +885,7 @@ describe('when given the load dkim guidance tag connection function', () => {
               expect(consoleWarnOutput).toEqual([
                 `User: ${
                   user._key
-                } attempted to have \`first\` set as a ${typeof invalidInput} for: dkimGuidanceTagConnectionsLoader.`,
+                } attempted to have \`first\` set as a ${typeof invalidInput} for: sslGuidanceTagConnectionsLoader.`,
               ])
             })
           })
@@ -895,21 +895,21 @@ describe('when given the load dkim guidance tag connection function', () => {
             it(`returns an error when last set to ${stringify(
               invalidInput,
             )}`, async () => {
-              const connectionLoader = dkimGuidanceTagConnectionsLoader(
+              const connectionLoader = sslGuidanceTagConnectionsLoader(
                 query,
                 user._key,
                 cleanseInput,
                 i18n,
               )
 
-              const dkimGuidanceTags = ['dkim1', 'dkim2']
+              const sslGuidanceTags = ['ssl1', 'ssl2']
               const connectionArgs = {
                 last: invalidInput,
               }
 
               try {
                 await connectionLoader({
-                  dkimGuidanceTags,
+                  sslGuidanceTags,
                   ...connectionArgs,
                 })
               } catch (err) {
@@ -918,7 +918,7 @@ describe('when given the load dkim guidance tag connection function', () => {
               expect(consoleWarnOutput).toEqual([
                 `User: ${
                   user._key
-                } attempted to have \`last\` set as a ${typeof invalidInput} for: dkimGuidanceTagConnectionsLoader.`,
+                } attempted to have \`last\` set as a ${typeof invalidInput} for: sslGuidanceTagConnectionsLoader.`,
               ])
             })
           })
@@ -931,20 +931,20 @@ describe('when given the load dkim guidance tag connection function', () => {
           .fn()
           .mockRejectedValue(new Error('Database Error Occurred.'))
 
-        const connectionLoader = dkimGuidanceTagConnectionsLoader(
+        const connectionLoader = sslGuidanceTagConnectionsLoader(
           query,
           user._key,
           cleanseInput,
           i18n,
         )
 
-        const dkimGuidanceTags = ['dkim1', 'dkim2']
+        const sslGuidanceTags = ['ssl1', 'ssl2']
         const connectionArgs = {
           first: 5,
         }
         try {
           await connectionLoader({
-            dkimGuidanceTags,
+            sslGuidanceTags,
             ...connectionArgs,
           })
         } catch (err) {
@@ -952,7 +952,7 @@ describe('when given the load dkim guidance tag connection function', () => {
         }
 
         expect(consoleErrorOutput).toEqual([
-          `Database error occurred while user: ${user._key} was trying to gather orgs in dkimGuidanceTagConnectionsLoader, error: Error: Database Error Occurred.`,
+          `Database error occurred while user: ${user._key} was trying to gather orgs in sslGuidanceTagConnectionsLoader, error: Error: Database Error Occurred.`,
         ])
       })
     })
@@ -965,20 +965,20 @@ describe('when given the load dkim guidance tag connection function', () => {
         }
         const query = jest.fn().mockReturnValueOnce(cursor)
 
-        const connectionLoader = dkimGuidanceTagConnectionsLoader(
+        const connectionLoader = sslGuidanceTagConnectionsLoader(
           query,
           user._key,
           cleanseInput,
           i18n,
         )
 
-        const dkimGuidanceTags = ['dkim1', 'dkim2']
+        const sslGuidanceTags = ['ssl1', 'ssl2']
         const connectionArgs = {
           first: 5,
         }
         try {
           await connectionLoader({
-            dkimGuidanceTags,
+            sslGuidanceTags,
             ...connectionArgs,
           })
         } catch (err) {
@@ -986,7 +986,7 @@ describe('when given the load dkim guidance tag connection function', () => {
         }
 
         expect(consoleErrorOutput).toEqual([
-          `Cursor error occurred while user: ${user._key} was trying to gather orgs in dkimGuidanceTagConnectionsLoader, error: Error: Cursor Error Occurred.`,
+          `Cursor error occurred while user: ${user._key} was trying to gather orgs in sslGuidanceTagConnectionsLoader, error: Error: Cursor Error Occurred.`,
         ])
       })
     })
